Crate xdr_codec

XDR runtime encoding/decoding

This crate provides runtime support for encoding and decoding XDR data. It is intended to be used with code generated by the "xdrgen" crate, but it can also be used with hand-written code.

It provides two key traits - Pack and Unpack - which all encodable types must implement. It also provides the helper functions pack() and unpack() to simplify the API.

Modules

record
XDR record marking

Enums

Error
XDR errors

Traits

Pack
Basic packing trait.

Read
The Read trait allows for reading bytes from a source.

Unpack
Basic unpacking trait

Write
A trait for objects which are byte-oriented sinks.

Functions

pack
Serialization (packing) helper.

pack_array
Pack a fixed-size array.

unpack
Deserialization (unpacking) helper function

unpack_flex_array
Unpack a length-limited array

unpack_string
Unpack length-limited string

Type Definitions

Result
A wrapper around std::result::Result where errors are all xdr_codec::Error.