Struct xdr_codec::record::XdrRecordReader

pub struct XdrRecordReader<R: BufRead> {
    // some fields omitted
}

Read records from a bytestream.

Reads will read up to the end of the current fragment, and not beyond. The BufRead trait doesn't otherwise allow for record boundaries to be deliniated. Callers can use the eor method to determine record ends.

Methods

impl<R: BufRead> XdrRecordReader<R>

fn new(rd: R) -> XdrRecordReader<R>

Wrapper a record reader around an existing implementation of BufRead, such as BufReader.

fn eor(&self) -> bool

Current fragment is the end of the record.