Struct xdr_codec::record::XdrRecordWriter

pub struct XdrRecordWriter<W: Write> {
    // some fields omitted
}

Write records into a bytestream.

Flushes the current buffer as end of record when destroyed.

Methods

impl<W: Write> XdrRecordWriter<W>

fn new(w: W) -> XdrRecordWriter<W>

Create a new XdrRecordWriter wrapped around a Write implementation, using a default buffer size (64k).

fn with_buffer(w: W, bufsz: usize) -> XdrRecordWriter<W>

Create an instance with a specific buffer size. Panics if the size is zero.

fn flush_eor(&mut self, eor: bool) -> Result<()>

Flush the current buffer. If eor is true, the end of record marker is set.